% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqtv.R
\name{calculate_cnvs}
\alias{calculate_cnvs}
\title{Calculate CNVs}
\usage{
calculate_cnvs(count_table)
}
\arguments{
\item{count_table}{A table of windowed counts read directly from the output of samtools depth. See worked example.}
}
\value{
A table of windows in which the coverage between the control and test sample(s) are compared.
}
\description{
Calculate CNVs
}
