% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqtv.R
\name{cnv_plot}
\alias{cnv_plot}
\title{Create a ggplot figure.}
\usage{
cnv_plot(
  cnv_table,
  hits,
  log2fc_thresh = 1.5,
  pval_thresh = 1e-12,
  zoom_chrom = NULL,
  zoom_start = NULL,
  zoom_end = NULL
)
}
\arguments{
\item{cnv_table}{Windowed comparisons between control and test sample(s), output directly from \code{calculate_cnvs}.}

\item{hits}{Hit ROIs from cnv_rois}

\item{log2fc_thresh}{log2 fold-change threshold value, keep windows with abs(log2_fc) > X}

\item{pval_thresh}{P-value threshold for signficant difference between test & control, windows with p < X are kept.}

\item{zoom_chrom}{Zoom to chrom}

\item{zoom_start}{Start position for zoom-in}

\item{zoom_end}{End position for zoom-in}
}
\value{
ROIs of merged genome windows passing these filtering thresholds
}
\description{
Create a ggplot figure.
}
