% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqtv.R
\name{cnv_rois}
\alias{cnv_rois}
\title{Extract ROIs from CNV calculation}
\usage{
cnv_rois(cnv_table, log2fc_thresh = 1.5, pval_thresh = 1e-12)
}
\arguments{
\item{cnv_table}{Windowed comparisons between control and test sample(s), output directly from \code{calculate_cnvs}.}

\item{log2fc_thresh}{log2 fold-change threshold value, keep windows with abs(log2_fc) > X}

\item{pval_thresh}{P-value threshold for signficant difference between test & control, windows with p < X are kept.}
}
\value{
ROIs of merged genome windows passing these filtering thresholds
}
\description{
Extract ROIs from CNV calculation
}
