% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqTV-package.R
\docType{package}
\name{cnvseqTV-package}
\alias{cnvseqTV}
\alias{cnvseqTV-package}
\title{cnvseqTV: Copy-number Variant detection with the tidyverse}
\description{
Detects large deletions or duplications as a result of mutagenesis from control and mutant samples.
}
\author{
\strong{Maintainer}: Kevin Murray \email{kevin.murray@gekkonid.com} (\href{https://orcid.org/0000-0002-2466-1917}{ORCID})

}
\keyword{internal}
