% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqtv.R
\name{ghratiop}
\alias{ghratiop}
\title{Calculate the Gearey-Hinkley transform p-value.}
\usage{
ghratiop(rat, xmean, ymean)
}
\arguments{
\item{rat}{Ratio between test and control}

\item{xmean}{Mean coverage of test}

\item{ymean}{Mean coverage of control}
}
\value{
Vector of p-values
}
\description{
Calculate the Gearey-Hinkley transform p-value.
}
