% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvseqtv.R
\name{overlap_groups}
\alias{overlap_groups}
\title{Overlap adjacent windows to define regions of interest}
\usage{
overlap_groups(starts, ends)
}
\arguments{
\item{starts}{Vector of start positions}

\item{ends}{Vector of end positions}
}
\value{
Vector of group identities
}
\description{
Overlap adjacent windows to define regions of interest
}
